mesh/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiohttp>=3.9.1
# Caching and utilities
cachetools>=5.3.2

# Search functionality
duckduckgo-search>=4.1.1

# Type hints support
typing-extensions>=4.9.0

# Async support
asyncio>=3.4.3

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2


openai>=1.40.8
firecrawl>=1.12.0
